requirements.txt
asgiref==3.8.1
click==8.1.7
dj-database-url==2.3.0
Django==5.0
django-cors-headers==4.5.0
django-heroku==0.3.1
django-nested-admin==4.1.1
djangorestframework==3.15.2
gunicorn==20.0.4
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2==2.9.10
Pygments==2.18.0
python-dotenv==1.0.1
python-monkey-business==1.1.0
pytz==2024.2
rich==13.9.4
shellingham==1.5.4
sqlparse==0.5.1
typer==0.13.0
typing_extensions==4.12.2
whitenoise==6.8.2
Pillow==11.0.0
openai

